Fix camera preview stuck on clicking app overlay button and reopen
When user reopens the aic app with previous app session in camera
preview state, camera preview is stuck.

When user clicks the overlay back button on the aic app, running apps
are not sent any close session or not put in background. So, even when
user clicks the overlay back button during active camera session
camera session is not closed resulting in camera session not cleaned
up properly.

Fix the issue by sending home key event when user clicks the overlay
button which is also inline with the behavior seen with chrome browser
client application.

Tracked-On: OAM-105261
Signed-off-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>